djangoproject/requirements.txt
Django>=3.2,<4.0
mysqlclient
djangorestframework
drf-yasg
djangorestframework-simplejwt
python-dotenv